IBoxData: Change money type to int

+++ ProjectShowoff/Assets/Scripts/Box/Data/IBoxData.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 
 public interface IBoxData<Contained>
 {
-	float MoneyValue { get; }
+	int MoneyValue { get; }
 	List<Contained> Contents { get; }
 
 	void AddToBox(Contained contained);
